pyproject.toml
[project]
name = "commons"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.4",
    "psycopg[binary]>=3.2.3",
    "pydantic-settings>=2.6.1",
    "sqlalchemy>=2.0.36",
]

[tool.uv]
dev-dependencies = [
    "coverage>=7.6.4",
    "faker>=30.8.2",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
    "ruff>=0.7.3",
]

[tool.ruff]
output-format = "concise"

[tool.ruff.lint]
select = [
    "E", "W",  # pycodestyle
    "F",  # Pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",  # isort
    "PL"  # Pylint
]

[tool.mypy]
strict = true
